intial commit

+++ TwitApp Sprint 2/TwitAppApi/TwitAppApi/Models/TblLogin.cs
@@ -14,6 +14,6 @@ namespace TwitAppApi.Models
         public string LoginId { get; set; }
         public string Password { get; set; }
         public string ConfirmPassword { get; set; }
-        public decimal? ContactNumber { get; set; }
+        public int? ContactNumber { get; set; }
     }
 }
